norminette

+++ src/exec/minimap/m_map_border.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   m_map_border.c                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: ofilloux <ofilloux@student.42barcelona.    +#+  +:+       +#+        */
+/*   By: ofilloux <ofilloux@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/08/27 22:58:21 by ofilloux          #+#    #+#             */
-/*   Updated: 2025/08/27 22:59:13 by ofilloux         ###   ########.fr       */
+/*   Updated: 2025/09/07 14:05:25 by ofilloux         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,8 +17,8 @@ void	draw_borders(t_global *data, t_mmap mmap, int col, int row)
 	int	outer_right;
 	int	outer_bottom;
 
-	outer_right = M_MAP_OFFSET_X + M_MAP_SIZE + 2 * M_MAP_BORDER;
-	outer_bottom = M_MAP_OFFSET_Y + M_MAP_SIZE + 2 * M_MAP_BORDER;
+	outer_right = M_MAP_OFFSET_X + (int)M_MAP_SIZE + 2 * M_MAP_BORDER;
+	outer_bottom = M_MAP_OFFSET_Y + (int)M_MAP_SIZE + 2 * M_MAP_BORDER;
 	if (col < M_MAP_OFFSET_X \
 		|| col > outer_right \
 		|| row < M_MAP_OFFSET_Y \
